+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG for Sulu
     * BUGFIX      #2925 [WebsiteBundle]       Fixed seo when no data is available
     * FEATURE     #2920 [WebsiteBundle]       Seo info as twig template to make it rewriteable
     * ENHANCEMENT #2915 [WebsiteBundle]       Refactored xml sitemap
+    * BUGFIX      #2923 [MediaBundle]         Fixes system collection creation for anon. authenticated users
     * BUGFIX      #2915 [MediaBundle]         Fixed height of badges in media-selection
     * ENHANCEMENT #2860 [ContentBundle]       Added url to internal links and smart-content
     * FEATURE     #2914 [CategoryBundle]      Added description and medias to category