+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG for Sulu
 ==================
 
 * dev-develop
+    * BUGFIX      #2797 [ContentBundle]       When removing resource locator in history now asks for conformation
     * BUGFIX      #2900 [ContentBundle]       Improved bug with grid elements not floating correctly
     * ENHANCEMENT #2927 [ContactBundle]       Enables the extensibility of matchings for contact-selection
     * BUGFIX      #2925 [WebsiteBundle]       Fixed seo when no data is available